feat: Добавил миграцию базы данных для каждого микросервиса через Flyway. А так же добавил Swagger UI для работы с ними.
fix: исправил Entity объекты для работы с flyway.А так же фикс остальных багов/ошибок.
