+++ .idea/compiler.xml
@@ -1,15 +1,25 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <annotationProcessing>
-      <profile name="Gradle Imported" enabled="true">
-        <outputRelativeToContentRoot value="true" />
-        <processorPath useClasspath="false">
-          <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.springframework.boot/spring-boot-configuration-processor/3.4.4/5103fbb3f4591f75c47be2f325a1d19a9dcf5804/spring-boot-configuration-processor-3.4.4.jar" />
-          <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.36/5a30490a6e14977d97d9c73c924c1f1b5311ea95/lombok-1.18.36.jar" />
-        </processorPath>
-        <module name="planner-users.main" />
-      </profile>
-    </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="planner-config" target="17" />
+      <module name="planner-entity" target="17" />
+      <module name="planner-microservice.planner-todo" target="17" />
+      <module name="planner-microservice.planner-todo.main" target="17" />
+      <module name="planner-microservice.planner-todo.test" target="17" />
+      <module name="planner-microservice.planner-users" target="17" />
+      <module name="planner-microservice.planner-users.main" target="17" />
+      <module name="planner-microservice.planner-users.test" target="17" />
+      <module name="resenkov.work.planner-config.main" target="17" />
+      <module name="resenkov.work.planner-config.test" target="17" />
+      <module name="resenkov.work.planner-entity.main" target="17" />
+      <module name="resenkov.work.planner-entity.test" target="17" />
+      <module name="resenkov.work.planner-gateway" target="17" />
+      <module name="resenkov.work.planner-gateway.main" target="17" />
+      <module name="resenkov.work.planner-gateway.test" target="17" />
+      <module name="resenkov.work.planner-server" target="17" />
+      <module name="resenkov.work.planner-server.main" target="17" />
+      <module name="resenkov.work.planner-server.test" target="17" />
+    </bytecodeTargetLevel>
   </component>
 </project>
