--- planner-entity/src/main/java/resenkov/work/plannerentity/entity/Priority.java
@@ -1,48 +1,0 @@
-package resenkov.work.plannerentity.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CacheConcurrencyStrategy;
-
-import java.util.Objects;
-
-@Getter
-@Setter
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "priority", schema = "todo",catalog = "planner_todo")
-@Cacheable
-@org.hibernate.annotations.Cache(usage = CacheConcurrencyStrategy.READ_WRITE)
-public class Priority {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "title", nullable = false, length = Integer.MAX_VALUE)
-    private String title;
-
-    @Column(name = "color", nullable = false, length = Integer.MAX_VALUE)
-    private String color;
-
-    @Column(name="user_id")
-    private Long userId;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Priority priority = (Priority) o;
-        return Objects.equals(id, priority.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-}
