+++ planner-todo/build.gradle
@@ -33,7 +33,8 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	compileOnly 'org.projectlombok:lombok'
 	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-aop', version: '3.4.0'
-	implementation('resenkov.work:planner-entity')
+	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.4.0'
+	implementation("org.flywaydb:flyway-core:9.16.0")
 }
 
 dependencyManagement {
