+++ planner-todo/src/main/java/resenkov/work/plannertodo/config/SwaggerConfig.java
@@ -1,0 +1,27 @@
+package resenkov.work.plannertodo.config;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import org.springdoc.core.models.GroupedOpenApi;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@OpenAPIDefinition(
+        info = @Info(
+                title = "Планировщик задач - работа с задачами",
+                version = "1.0",
+                description = "API для работы с задачами"
+        )
+)
+public class SwaggerConfig {
+
+    @Bean
+    public GroupedOpenApi allApi() {
+        return GroupedOpenApi.builder()
+                .group("all-apis")
+                .pathsToMatch("/**")
+                .build();
+    }
+
+}
