+++ planner-todo/src/main/java/resenkov/work/plannertodo/controller/StatController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import resenkov.work.plannerentity.entity.Stat;
+import resenkov.work.plannertodo.entity.Stat;
 import resenkov.work.plannertodo.service.StatService;
 
 @RestController
