+++ planner-todo/src/main/java/resenkov/work/plannertodo/repo/PriorityRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import resenkov.work.plannerentity.entity.Priority;
+import resenkov.work.plannertodo.entity.Priority;
+
 
 import java.util.List;
