+++ planner-users/src/main/java/resenkov/work/plannerusers/config/SwaggerConfig.java
@@ -1,0 +1,27 @@
+package resenkov.work.plannerusers.config;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import org.springdoc.core.models.GroupedOpenApi;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@OpenAPIDefinition(
+        info = @Info(
+                title = "Планировщик задач - пользовательская часть",
+                version = "1.0",
+                description = "API для работы с пользователями"
+        )
+)
+public class SwaggerConfig {
+
+    @Bean
+    public GroupedOpenApi allApi() {
+        return GroupedOpenApi.builder()
+                .group("all-apis")
+                .pathsToMatch("/**")
+                .build();
+    }
+
+}
