+++ planner-users/src/main/java/resenkov/work/plannerusers/repository/UserRepository.java
@@ -3,8 +3,7 @@ package resenkov.work.plannerusers.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import resenkov.work.plannerentity.entity.User;
-
+import resenkov.work.plannerusers.entity.User;
 
 
 @Repository
