+++ planner-users/src/main/resources/application.properties
@@ -5,3 +5,13 @@ spring.config.import=configserver:http://localhost:8888
 
 
 spring.profiles.active=micro, user, logging
+
+springdoc.swagger-ui.path=/swagger-ui.html
+springdoc.api-docs.path=/api-docs
+springdoc.swagger-ui.enabled=true
+springdoc.api-docs.enabled=true
+
+spring.flyway.enabled=true
+spring.flyway.locations=classpath:db/migration
+spring.flyway.baseline-on-migrate=true
+spring.flyway.schemas=public
