+++ settings.gradle
@@ -1,7 +1,8 @@
 rootProject.name = 'planner-microservice'
+
 includeBuild('planner-server')
 includeBuild('planner-gateway')
-includeBuild('planner-config')
-includeBuild('planner-entity')
+includeBuild ('planner-config')
+includeBuild ('planner-entity')
 includeBuild('planner-todo')
 includeBuild('planner-users')
